+++ flow-server/src/main/java/com/vaadin/flow/component/internal/StyleSheetHotswapper.java
@@ -46,11 +46,11 @@ import com.vaadin.flow.hotswap.UIUpdateStrategy;
 import com.vaadin.flow.hotswap.VaadinHotswapper;
 import com.vaadin.flow.internal.ActiveStyleSheetTracker;
 import com.vaadin.flow.internal.AnnotationReader;
+import com.vaadin.flow.internal.CssBundler;
 import com.vaadin.flow.internal.FrontendUtils;
 import com.vaadin.flow.server.AppShellRegistry;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.server.VaadinSession;
-import com.vaadin.flow.server.frontend.CssBundler;
 import com.vaadin.flow.server.startup.ApplicationConfiguration;
 import com.vaadin.flow.shared.ApplicationConstants;
 import com.vaadin.flow.shared.ui.Dependency;
