+++ flow-server/src/main/java/com/vaadin/flow/internal/ThemeUtils.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.server.frontend;
+package com.vaadin.flow.internal;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,9 +33,6 @@ import org.slf4j.LoggerFactory;
 import tools.jackson.databind.JsonNode;
 
 import com.vaadin.flow.component.page.AppShellConfigurator;
-import com.vaadin.flow.internal.FileIOUtils;
-import com.vaadin.flow.internal.FrontendUtils;
-import com.vaadin.flow.internal.JacksonUtils;
 import com.vaadin.flow.server.AbstractConfiguration;
 import com.vaadin.flow.server.AppShellRegistry;
 import com.vaadin.flow.server.Constants;
