+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/ThemeValidationUtil.java
@@ -42,6 +42,7 @@ import tools.jackson.databind.node.JsonNodeType;
 import com.vaadin.flow.internal.FileIOUtils;
 import com.vaadin.flow.internal.JacksonUtils;
 import com.vaadin.flow.internal.StringUtil;
+import com.vaadin.flow.internal.ThemeUtils;
 import com.vaadin.flow.server.Constants;
 import com.vaadin.flow.server.frontend.scanner.FrontendDependenciesScanner;
 import com.vaadin.flow.theme.ThemeDefinition;
