+++ vaadin-dev-server/src/main/java/com/vaadin/base/devserver/PublicStyleSheetBundler.java
@@ -24,8 +24,8 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.vaadin.flow.internal.CssBundler;
 import com.vaadin.flow.internal.FrontendUtils;
-import com.vaadin.flow.server.frontend.CssBundler;
 import com.vaadin.flow.shared.ApplicationConstants;
 
 /**
