+++ vaadin-dev-server/src/main/java/com/vaadin/base/devserver/ThemeLiveUpdater.java
@@ -26,9 +26,9 @@ import tools.jackson.databind.JsonNode;
 
 import com.vaadin.flow.internal.BrowserLiveReload;
 import com.vaadin.flow.internal.BrowserLiveReloadAccessor;
+import com.vaadin.flow.internal.CssBundler;
+import com.vaadin.flow.internal.ThemeUtils;
 import com.vaadin.flow.server.VaadinContext;
-import com.vaadin.flow.server.frontend.CssBundler;
-import com.vaadin.flow.server.frontend.ThemeUtils;
 import com.vaadin.flow.server.startup.ApplicationConfiguration;
 
 /**
